Agrego logs, observabilidad, y manejo de errores

+++ src/main/java/ar/utn/ba/ddsi/apiadmi/servicies/ColeccionesServices.java
@@ -1,5 +1,6 @@
 package ar.utn.ba.ddsi.apiadmi.servicies;
 
+import ar.utn.ba.ddsi.apiadmi.exception.RecursoNoEncontradoException;
 import ar.utn.ba.ddsi.apiadmi.models.dtos.ColeccionDto;
 import ar.utn.ba.ddsi.apiadmi.models.dtos.CondicionDTO;
 import ar.utn.ba.ddsi.apiadmi.models.dtos.input.ColeccionInput;
@@ -14,6 +15,7 @@ import ar.utn.ba.ddsi.apiadmi.models.repository.IColeccionRepository;
 import ar.utn.ba.ddsi.apiadmi.servicies.interfaces.IColeccionService;
 import ar.utn.ba.ddsi.apiadmi.servicies.interfaces.IFuenteServices;
 import ar.utn.ba.ddsi.apiadmi.utils.EnumTipoDeAlgoritmo;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,7 +26,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import static java.lang.Long.parseLong;
-
+@Slf4j
 
 @Service
 public class ColeccionesServices implements IColeccionService {
@@ -49,11 +51,23 @@ public class ColeccionesServices implements IColeccionService {
 
 
     public List<ColeccionDto> obtenerColecciones() {
-        return this.colecciones.findAll().stream().map(this::ColeccionDto).toList();
-        //if(this.colecciones.findAll().stream().map(this::ColeccionDto).toList();)
+
+        log.info("Obteniendo todas las colecciones");
+        List<Coleccion> colecciones = this.colecciones.findAll();
+        if(colecciones.isEmpty()){
+            log.warn("No se encontraron colecciones en la base de datos");
+        } else {
+            log.debug("Colecciones obtenidas: {}", colecciones.size());
+            return this.colecciones.findAll().stream().map(this::ColeccionDto).toList();
+        }
+        return new ArrayList<>();
+
+
     }
 /* PARA MOSTRAR TODAS LAS COLECCIONES EN GENERAL*/
     private ColeccionDto ColeccionDto(Coleccion cole) {
+        log.debug("Convirtiendo colección a DTO: {}", cole.getTitulo());
+
         ColeccionDto coleout = new ColeccionDto();
         coleout.setId_coleccion(cole.getId_coleccion());
         coleout.setTitulo(cole.getTitulo());
@@ -96,14 +110,22 @@ public class ColeccionesServices implements IColeccionService {
     @Override
     public void agregar(ColeccionInput coleccion) {
 
-        List<Fuente> fuentes = coleccion.getFuentes().stream()
-                .map(a -> this.fuenteService.buscarPorNombre(a))
-                .collect(Collectors.toList());
+        log.info("Iniciando creación de colección: {}",
+                coleccion.getTitulo());
 
+        List<Fuente> fuentes = coleccion.getFuentes().stream().map(nombre -> {
+            log.debug("Buscando fuente: {}", nombre);
+            return fuenteService.buscarPorNombre(nombre);
+        }).toList();
+
+        log.debug("Fuentes encontradas: {}", fuentes.size());
+
+        log.info("Creando o relacionando criterios para la colección: {}" ,coleccion.getTitulo());
         List<InterfaceCondicion> criterios = coleccion.getCriterios().stream()
                 .map(a-> this.cargarOCrearCondicion(a))
                 .collect(Collectors.toList());
 
+        log.debug("Condicones relacionadas: {}", criterios.size());
 
         Coleccion cole = this.coleccionFactory.crearColeccion(coleccion);
         cole.setFuentes(fuentes);
@@ -112,22 +134,32 @@ public class ColeccionesServices implements IColeccionService {
 
 
         colecciones.save(cole);
+        log.info("Colección creada correctamente");
     }
 
     @Transactional
     @Override
     public void actualizar(Long id,ColeccionInput input){
 
+
+        log.info("Actualizando colección id={}", id);
+
         try {
+
 
             Coleccion cole = colecciones.findById(id)
                     .orElseThrow(() -> new RuntimeException("No existe la colección"));
             System.out.println(input);
+
+            log.debug("Coleccion encontrada: {}", id);
+
             cole.setTitulo(input.getTitulo());
             cole.setDescripcion(input.getDescripcion());
+
             //Actualizacion de algoritmo de concenso
             String algoInput = input.getAlgoritmoConcenso();
             if(algoInput != null) {
+                log.info("Actualizando algoritmo de consenso a {} para colección id={}", algoInput, id);
             cole.setTipoDeAlgoritmo(EnumTipoDeAlgoritmo.valueOf(input.getAlgoritmoConcenso()));
             }
             // fuentes
@@ -135,6 +167,7 @@ public class ColeccionesServices implements IColeccionService {
                     .map(f -> fuenteService.buscarPorNombre(f))
                     .collect(Collectors.toList());
             cole.setFuentes(nuevasFuentes);
+            log.debug("Fuentes actualizadas para colección id={}: {}", id, nuevasFuentes.size());
 
             List<InterfaceCondicion> condicionesOriginales = cole.getCondicionDePertenencia();
 
@@ -156,35 +189,50 @@ public class ColeccionesServices implements IColeccionService {
             cole.getCondicionDePertenencia().clear();
 
             // Agregar las nuevas
+            log.debug("Condiciones actualizadas de coleccion id={}: {}", id, nuevasCondiciones.size());
             cole.getCondicionDePertenencia().addAll(nuevasCondiciones);
             
             colecciones.save(cole);
         } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
+            log.error("Error actualizando colección id={}", id, e);
 
+            throw e;
+        }
     }
 
     public InterfaceCondicion cargarOCrearCondicion(CondicionInput inputCOndicion){
+
+        log.debug("Procesando condición tipo={}",inputCOndicion.getTipo());
+
         if (inputCOndicion.getId() != null) {
             return condicionService.buscarPorId(inputCOndicion.getId());
         }
         // la creo según el tipo
+        log.debug("Creando nueva condición del tipo {} con valor {}", inputCOndicion.getTipo(), inputCOndicion.getValor());
         return this.crearCondicion(inputCOndicion);
     }
 
     @Override
     public Coleccion encontrarPorId(Long id) {
 
+        log.info("Obteniendo la colección con id={}", id);
         Coleccion c = colecciones.findById(id)
-                .orElseThrow(() -> new RuntimeException("Colección no encontrada"));
+                .orElseThrow(() -> {
+                    log.error("Colección no encontrada con id={}", id);
+                    return new RuntimeException("Colección no encontrada");});
 
         return c;
     }
 
     @Override
     public void eliminar(Long id) {
-        this.colecciones.deleteById(id);
+            log.info("Eliminando colección con id={}", id);
+            if (!colecciones.existsById(id)) {
+                log.error("No se encontró la colección con id={}", id);
+                throw new RuntimeException("Colección no encontrada");
+            }
+            log.debug("Colección encontrada, procediendo a eliminar id={}", id);
+            this.colecciones.deleteById(id);
 
 
     }
@@ -212,15 +260,41 @@ public class ColeccionesServices implements IColeccionService {
                 return new CondicionFechaANTES(hasta);
 
             case "categoria":
-                Categoria categoria= this.categoriaService.buscarPorNombre(valor);
-                if(categoria!=null) return new CondicionCategoria(categoria);
+
+                Categoria categoria =
+                        categoriaService.buscarPorNombre(valor);
+
+                if (categoria == null) {
+
+                    log.warn("Categoría no encontrada: {}", valor);
+
+                    throw new RecursoNoEncontradoException(
+                            "Categoría no encontrada: " + valor);
+                }
+
+                return new CondicionCategoria(categoria);
+
             case "etiqueta":
                 Etiqueta etiqueta =this.etiquetaService.buscarPorId(valor);
-                if(etiqueta!=null) return new CondicionEtiqueta(etiqueta);
+                if(etiqueta==null) {
+                    log.warn("Etiqueta no encontrada: {}", valor);
+
+                    throw new RecursoNoEncontradoException(
+                            "Etiqueta no encontrada: " + valor);
+                }
+                return new CondicionEtiqueta(etiqueta);
 
             case "fuente":
                     Fuente fuente = this.fuenteService.buscarPorId(parseLong(valor));
-                    if(fuente!=null) return new CondicionFuente(fuente);
+
+                    if(fuente==null){
+                        log.warn("Fuente no encontrada: {}", valor);
+
+                        throw new RecursoNoEncontradoException(
+                                "Fuente no encontrada: " + valor);
+                    }
+
+                    return new CondicionFuente(fuente);
             default :
                 throw new IllegalArgumentException("Tipo de condición no soportado: " + tipo);
         }
